mechanism to display multiple pictures
tldr:
imageImport.js will be ran everytime you runs the app.

generate an imageImport.js in the src folder. This files is the map to all images found in the images folder. Allows us to display pictures dynamically since react ntive HATE dynamic images.

realistically, this wont be able to work as a social media app. But as a demo, yesssir